rename `QueryInfo` to `MergeStrategy`, move into `cache/inmemory`
