+++ src/core/index.ts
@@ -42,7 +42,6 @@ export type {
   MutationUpdaterFunction,
   OnQueryUpdated,
   OperationVariables,
-  QueryListener,
   QueryResult,
   RefetchQueriesInclude,
   RefetchQueriesOptions,
