+++ src/core/types.ts
@@ -14,7 +14,6 @@ import type { IsAny } from "@apollo/client/utilities/internal";
 
 import type { NetworkStatus } from "./networkStatus.js";
 import type { ObservableQuery } from "./ObservableQuery.js";
-import type { QueryInfo } from "./QueryInfo.js";
 import type { QueryOptions } from "./watchQueryOptions.js";
 
 export type { TypedDocumentNode } from "@graphql-typed-document-node/core";
@@ -37,8 +36,6 @@ export interface DefaultContext extends Record<string, any> {
   queryDeduplication?: boolean;
   clientAwareness?: ClientAwareness;
 }
-
-export type QueryListener = (queryInfo: QueryInfo) => void;
 
 /**
  * Represents an `Error` type, but used throughout Apollo Client to represent
